+++ packages/client/src/StreamrClient.ts
@@ -136,14 +136,13 @@ export class StreamrClient {
     }
 
     /**
-     * Adds an encryption key for a given stream to the key store.
+     * Adds an encryption key for a given publisher to the key store.
      *
      * @remarks Keys will be added to the store automatically by the client as encountered. This method can be used to
      * manually add some known keys into the store.
      */
-    async addEncryptionKey(key: GroupKey, streamIdOrPath: string): Promise<void> {
-        const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
-        await this.groupKeyStore.add(key, streamId)
+    async addEncryptionKey(key: GroupKey, publisherId: EthereumAddress): Promise<void> {
+        await this.groupKeyStore.add(key, publisherId)
     }
 
     // --------------------------------------------------------------------------------------------